refactor: Cars 객체에 initCars 역할 위임
