+++ src/main/java/racingcar/model/RacingGame.java
@@ -2,50 +2,22 @@ package racingcar.model;
 
 import racingcar.util.InputValidator;
 
-import java.util.Arrays;
-import java.util.List;
-import java.util.stream.Collectors;
-
 public class RacingGame {
-    private static final String COMMA = ",";
-    private static final String BLANK = "";
-    private static final String SPACE = " ";
-
     private Cars cars;
     private int trialNum;
 
     public RacingGame() {
     }
 
+    //TODO 생성자를 이렇게 사용해도 될까 고민해보기
     public RacingGame(String carNames, String inputTrialNum) {
-        List<String> splitCarNames = splitCars(carNames);
-        this.cars = initCars(splitCarNames);
+        this.cars = Cars.initCars(carNames);
         this.trialNum = convertToInteger(inputTrialNum);
     }
 
     private int convertToInteger(String number) {
         InputValidator.validateTrialInput(number);
         return Integer.parseInt(number);
-    }
-
-    public List<String> splitCars(String carNames) {
-        InputValidator.validateNameInput(carNames);
-        List<String> splitCarNames = Arrays.asList(splitNameWithoutSpace(carNames));
-        InputValidator.validateEmptyCarName(splitCarNames);
-        InputValidator.validateCarNameLength(splitCarNames);
-        InputValidator.validateDuplicateName(splitCarNames);
-
-        return splitCarNames;
-    }
-
-    private String[] splitNameWithoutSpace(String carNames) {
-        return carNames.replaceAll(SPACE, BLANK).split(COMMA);
-    }
-
-    public Cars initCars(List<String> carNames) {
-        return new Cars(carNames.stream()
-                .map(Car::new)
-                .collect(Collectors.toList()));
     }
 
     public void race() {
@@ -59,5 +31,4 @@ public class RacingGame {
     public int getTrialNum() {
         return trialNum;
     }
-
 }
